requirements.txt
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
